feat: implement PDF generation for U029 with structured document layout and improved data handling
