+++ src/main/kotlin/no/nav/eux/pdf/service/U020PdfGenerator.kt
@@ -1,17 +1,8 @@
 package no.nav.eux.pdf.service
 
-import io.github.oshai.kotlinlogging.KotlinLogging.logger
 import org.apache.pdfbox.pdmodel.PDDocument
-import org.apache.pdfbox.pdmodel.PDPage
-import org.apache.pdfbox.pdmodel.PDPageContentStream
-import org.apache.pdfbox.pdmodel.common.PDRectangle
-import org.apache.pdfbox.pdmodel.font.PDFont
-import org.apache.pdfbox.pdmodel.font.PDType0Font
 import java.io.ByteArrayOutputStream
 import java.io.IOException
-import java.time.LocalDate
-import java.time.LocalDateTime.now
-import java.time.format.DateTimeFormatter.ofPattern
 
 data class U020Master(
     val rinasakId: String,
@@ -58,39 +49,15 @@ data class U020Child(
     val requestedCurrency: String
 )
 
-data class PersonIdInfo(
-    val country: String,
-    val personalIdentificationNumber: String,
-    val sector: String,
-    val institutionID: String,
-    val institutionName: String
-)
-
-data class PlaceBirthInfo(
-    val town: String? = null,
-    val region: String? = null,
-    val country: String? = null
-)
-
 class EessiU020PdfGen {
-
-    val pageWidth = PDRectangle.A4.width
-    val pageHeight = PDRectangle.A4.height
-    val marginLeft = 50f
-    val marginRight = 50f
-    val marginTop = 40f
-    val marginBottom = 50f
-    val lineHeight = 18f
-
-    private val log = logger {}
 
     fun generateU020Document(master: U020Master, claims: List<U020Child>): ByteArray =
         try {
             val document = PDDocument()
-            val writer = PdfWriter(document)
+            val writer = U020PdfWriter(document)
 
             writer.writeRinasakIdTopRight(master.rinasakId)
-            writer.writeDocumentTitle()
+            writer.writeDocumentTitle("U020 - Forespørsel om refusjon")
             writer.writeGeneratedDate()
             writer.addBlankLine()
 
@@ -109,157 +76,7 @@ class EessiU020PdfGen {
             throw RuntimeException("Failed to generate EESSI U020 PDF", e)
         }
 
-    private inner class PdfWriter(val document: PDDocument) {
-        var currentPage: PDPage = createNewPage()
-        var contentStream: PDPageContentStream = PDPageContentStream(document, currentPage)
-        var currentY: Float = pageHeight - marginTop
-        var pageNumber: Int = 1
-        val boldFont by lazy { loadFont("NotoSans-Bold.ttf") }
-        val regularFont by lazy { loadFont("NotoSans-Regular.ttf") }
-        val italicFont by lazy { loadFont("NotoSans-Italic.ttf") }
-
-        private fun createNewPage(): PDPage {
-            val page = PDPage(PDRectangle.A4)
-            document.addPage(page)
-            return page
-        }
-
-        private fun writeFooter() {
-            val footerY = marginBottom - 15f
-            val footerLineY = marginBottom - 5f
-
-            contentStream.beginText()
-            contentStream.setFont(regularFont, 9f)
-            contentStream.newLineAtOffset(marginLeft, footerY)
-            contentStream.showText("U020")
-            contentStream.endText()
-
-            val pageText = "Side $pageNumber"
-            val pageTextWidth = regularFont.getStringWidth(pageText) / 1000 * 9f
-            contentStream.beginText()
-            contentStream.setFont(regularFont, 9f)
-            contentStream.newLineAtOffset(pageWidth - marginRight - pageTextWidth, footerY)
-            contentStream.showText(pageText)
-            contentStream.endText()
-
-            contentStream.setLineWidth(0.5f)
-            contentStream.moveTo(marginLeft, footerLineY)
-            contentStream.lineTo(pageWidth - marginRight, footerLineY)
-            contentStream.stroke()
-        }
-
-        private fun checkPageSpace(requiredSpace: Float = lineHeight) {
-            if (currentY - requiredSpace < marginBottom + 30f) {
-                writeFooter()
-                contentStream.close()
-                currentPage = createNewPage()
-                contentStream = PDPageContentStream(document, currentPage)
-                currentY = pageHeight - marginTop
-                pageNumber++
-            }
-        }
-
-        fun writeDocumentTitle() {
-            checkPageSpace(40f)
-
-            contentStream.beginText()
-            contentStream.setFont(boldFont, 18f)
-            contentStream.newLineAtOffset(marginLeft, currentY)
-            contentStream.showText("U020 - Forespørsel om refusjon")
-            contentStream.endText()
-
-            currentY -= 15f
-
-            contentStream.setLineWidth(1f)
-            contentStream.setStrokingColor(0.7f, 0.7f, 0.7f)
-            contentStream.moveTo(marginLeft, currentY)
-            contentStream.lineTo(pageWidth - marginRight, currentY)
-            contentStream.stroke()
-
-            currentY -= 10f
-        }
-
-        fun writeGeneratedDate() {
-            checkPageSpace()
-            val dateString = "Generert: ${now().format(ofPattern("dd.MM.yyyy 'kl.' HH:mm"))}"
-            contentStream.beginText()
-            contentStream.setFont(italicFont, 8f)
-            contentStream.setNonStrokingColor(0.5f, 0.5f, 0.5f)
-            contentStream.newLineAtOffset(marginLeft, currentY)
-            contentStream.showText(dateString)
-            contentStream.endText()
-            contentStream.setNonStrokingColor(0f, 0f, 0f)
-            currentY -= lineHeight
-        }
-
-        fun writeSectionHeader(title: String) {
-            checkPageSpace(30f)
-            addBlankLine()
-
-            contentStream.beginText()
-            contentStream.setFont(boldFont, 12f)
-            contentStream.newLineAtOffset(marginLeft, currentY)
-            contentStream.showText(title)
-            contentStream.endText()
-            currentY -= 25f
-        }
-
-        fun writeSubsectionHeader(title: String) {
-            checkPageSpace(20f)
-            contentStream.beginText()
-            contentStream.setFont(boldFont, 10f)
-            contentStream.setNonStrokingColor(0.2f, 0.2f, 0.2f)
-            contentStream.newLineAtOffset(marginLeft + 10f, currentY)
-            contentStream.showText(title)
-            contentStream.endText()
-            contentStream.setNonStrokingColor(0f, 0f, 0f)
-            currentY -= 18f
-        }
-
-        fun writeKeyValuePair(key: String, value: String, indent: Float = 20f) {
-            checkPageSpace()
-            val keyText = "$key:"
-            val keyWidth = boldFont.getStringWidth(keyText) / 1000 * 12f
-
-            contentStream.beginText()
-            contentStream.setFont(boldFont, 12f)
-            contentStream.newLineAtOffset(marginLeft + indent, currentY)
-            contentStream.showText(keyText)
-            contentStream.endText()
-
-            contentStream.beginText()
-            contentStream.setFont(regularFont, 12f)
-            contentStream.newLineAtOffset(marginLeft + indent + keyWidth + 5f, currentY)
-            contentStream.showText(value)
-            contentStream.endText()
-
-            currentY -= lineHeight
-        }
-
-        fun writeCompactKeyValuePair(key: String, value: String, startX: Float) {
-            val keyText = "$key:"
-            val keyWidth = boldFont.getStringWidth(keyText) / 1000 * 12f
-
-            contentStream.beginText()
-            contentStream.setFont(boldFont, 12f)
-            contentStream.newLineAtOffset(startX, currentY)
-            contentStream.showText(keyText)
-            contentStream.endText()
-
-            contentStream.beginText()
-            contentStream.setFont(regularFont, 12f)
-            contentStream.newLineAtOffset(startX + keyWidth + 3f, currentY)
-            contentStream.showText(value)
-            contentStream.endText()
-        }
-
-        fun addBlankLine() {
-            currentY -= lineHeight * 0.8f
-        }
-
-        fun addSmallSpace() {
-            currentY -= lineHeight * 0.3f
-        }
+    private class U020PdfWriter(document: PDDocument) : BasePdfWriter(document, "U020") {
 
         fun writeMasterInformation(master: U020Master) {
             writeSectionHeader("Dokumentinformasjon")
@@ -300,17 +117,7 @@ class EessiU020PdfGen {
                 writeSingleClaim(claim, index, claims.size)
             }
 
-            writeFooter()
-            contentStream.close()
-        }
-
-        private fun startNewPageForClaims() {
-            writeFooter()
-            contentStream.close()
-            currentPage = createNewPage()
-            contentStream = PDPageContentStream(document, currentPage)
-            currentY = pageHeight - marginTop
-            pageNumber++
+            close()
         }
 
         private fun writeSingleClaim(claim: U020Child, index: Int, totalClaims: Int) {
@@ -333,18 +140,7 @@ class EessiU020PdfGen {
                 ?.takeIf { it.isNotEmpty() }
                 ?.let { it.size + 1 }
                 ?: 0
-            return 120f + ((baseLines + additionalLines) * lineHeight)
-        }
-
-        private fun ensureSufficientSpace(requiredSpace: Float) {
-            if (currentY - requiredSpace < marginBottom + 30f) {
-                writeFooter()
-                contentStream.close()
-                currentPage = createNewPage()
-                contentStream = PDPageContentStream(document, currentPage)
-                currentY = pageHeight - marginTop
-                pageNumber++
-            }
+            return 120f + ((baseLines + additionalLines) * 18f)
         }
 
         private fun writeClaimDetails(claim: U020Child) {
@@ -364,33 +160,8 @@ class EessiU020PdfGen {
             }
         }
 
-        private fun writeIdNumbersHeader() {
-            contentStream.beginText()
-            contentStream.setFont(boldFont, 12f)
-            contentStream.setNonStrokingColor(0f, 0f, 0f)
-            contentStream.newLineAtOffset(marginLeft + 30f, currentY)
-            contentStream.showText("ID-nummer:")
-            contentStream.endText()
-            currentY -= 16f
-        }
-
-        private fun writeIdNumber(pin: PersonIdInfo) {
-            val maskedPnr = maskNorwegianPnr(pin.personalIdentificationNumber, pin.country)
-            contentStream.beginText()
-            contentStream.setFont(regularFont, 12f)
-            contentStream.setNonStrokingColor(0.3f, 0.3f, 0.3f)
-            contentStream.newLineAtOffset(marginLeft + 45f, currentY)
-            contentStream.showText("${pin.country}: $maskedPnr (${pin.sector})")
-            contentStream.endText()
-            currentY -= 16f
-        }
-
-        private fun restoreDefaultTextColor() {
-            contentStream.setNonStrokingColor(0f, 0f, 0f)
-        }
-
         private fun writePersonalInformation(claim: U020Child) {
-            val columnX = marginLeft + 30f
+            val columnX = getMarginLeft() + 30f
 
             writeBasicPersonalInfo(claim, columnX)
             writeBirthInfo(claim, columnX)
@@ -430,16 +201,8 @@ class EessiU020PdfGen {
             }
         }
 
-        private fun formatPlaceOfBirth(placeBirth: PlaceBirthInfo?): String? =
-            placeBirth?.let { place ->
-                listOfNotNull(place.town, place.region, place.country)
-                    .filter { it.isNotBlank() }
-                    .joinToString(", ")
-                    .takeIf { it.isNotBlank() }
-            }
-
         private fun writeInstitutionalInformation(claim: U020Child) {
-            val columnX = marginLeft + 30f
+            val columnX = getMarginLeft() + 30f
 
             writeCompactKeyValuePair("Sekvensnr", claim.sequentialNumber, columnX)
             currentY -= 16f
@@ -459,7 +222,8 @@ class EessiU020PdfGen {
             writeCompactKeyValuePair("Siste utbetaling", formatDate(claim.lastPaymentDate), columnX)
             currentY -= 16f
 
-            val reimbursementPeriod = "${formatDate(claim.reimbursementPeriodStart)} - ${formatDate(claim.reimbursementPeriodEnd)}"
+            val reimbursementPeriod =
+                "${formatDate(claim.reimbursementPeriodStart)} - ${formatDate(claim.reimbursementPeriodEnd)}"
             writeCompactKeyValuePair("Refusjonsperiode", reimbursementPeriod, columnX)
             currentY -= 16f
         }
@@ -468,121 +232,5 @@ class EessiU020PdfGen {
             writeCompactKeyValuePair("Beløp", "${claim.requestedAmount} ${claim.requestedCurrency}", columnX)
             currentY -= 17f
         }
-
-        private fun shouldDrawSeparator(index: Int, totalClaims: Int): Boolean =
-            index < totalClaims - 1
-
-        private fun drawClaimSeparator() {
-            contentStream.setLineWidth(0.5f)
-            contentStream.setStrokingColor(0.9f, 0.9f, 0.9f)
-            contentStream.moveTo(marginLeft + 20f, currentY)
-            contentStream.lineTo(pageWidth - marginRight - 20f, currentY)
-            contentStream.stroke()
-            currentY -= 8f
-        }
-
-        fun writeRinasakIdTopRight(rinasakId: String) {
-            val text = "Saksnr: $rinasakId"
-            val textWidth = boldFont.getStringWidth(text) / 1000 * 8f
-
-            contentStream.beginText()
-            contentStream.setFont(boldFont, 8f)
-            contentStream.newLineAtOffset(pageWidth - marginRight - textWidth, pageHeight - marginTop - 10f)
-            contentStream.showText(text)
-            contentStream.endText()
-        }
-
-        private fun loadFont(fontFileName: String): PDFont {
-            val fontStream = javaClass.getResourceAsStream("/fonts/$fontFileName")
-                ?: throw IllegalStateException("Font file $fontFileName not found in resources/fonts/")
-            return PDType0Font.load(document, fontStream)
-        }
     }
-
-
-    fun formatDate(dateString: String): String =
-        try {
-            val date = LocalDate.parse(dateString)
-            date.format(ofPattern("dd.MM.yyyy"))
-        } catch (e: Exception) {
-            dateString
-        }
-
-    fun getSexDescription(sexCode: String): String =
-        when (sexCode) {
-            "01" -> "Mann"
-            "02" -> "Kvinne"
-            else -> "Ukjent ($sexCode)"
-        }
-
-    fun maskNorwegianPnr(pnr: String, country: String) =
-        when (country.uppercase()) {
-            "NO" -> {
-                if (pnr.length == 11 && pnr.all { it.isDigit() }) {
-                    val firstSix = pnr.substring(0, 6)
-                    val masked = "*".repeat(5)
-                    "$firstSix$masked"
-                } else {
-                    maskGeneric(pnr)
-                }
-            }
-
-            "DK" -> {
-                val cleanPnr = pnr.replace("-", "")
-                if (cleanPnr.length == 10 && cleanPnr.all { it.isDigit() }) {
-                    val birthDate = cleanPnr.substring(0, 6)
-                    val masked = "*".repeat(4)
-                    "$birthDate-$masked"
-                } else {
-                    maskGeneric(pnr)
-                }
-            }
-
-            "SE" -> {
-                val cleanPnr = pnr.replace("-", "")
-                when {
-                    cleanPnr.length == 12 && cleanPnr.all { it.isDigit() } -> {
-                        val birthDate = cleanPnr.substring(0, 8)
-                        val masked = "*".repeat(4)
-                        "$birthDate-$masked"
-                    }
-
-                    cleanPnr.length == 10 && cleanPnr.all { it.isDigit() } -> {
-                        val birthDate = cleanPnr.substring(0, 6)
-                        val masked = "*".repeat(4)
-                        "$birthDate-$masked"
-                    }
-
-                    else -> maskGeneric(pnr)
-                }
-            }
-
-            "FI" -> {
-                if (pnr.length == 11) {
-                    val birthDate = pnr.substring(0, 6)
-                    val centuryMarker = pnr[6]
-                    val validCenturyMarkers = setOf(
-                        '+', '-', 'Y', 'X', 'W', 'V', 'U', 'T', 'S', 'R', 'Q',
-                        'P', 'N', 'M', 'L', 'K', 'J', 'H', 'G', 'F', 'E', 'D', 'C', 'B', 'A'
-                    )
-
-                    if (birthDate.all { it.isDigit() } && centuryMarker in validCenturyMarkers) {
-                        val masked = "*".repeat(4)
-                        "$birthDate$centuryMarker$masked"
-                    } else {
-                        maskGeneric(pnr)
-                    }
-                } else {
-                    maskGeneric(pnr)
-                }
-            }
-
-            else -> maskGeneric(pnr)
-        }
-
-    private fun maskGeneric(pnr: String): String =
-        if (pnr.length > 3)
-            "*".repeat(pnr.length - 3) + pnr.takeLast(3)
-        else
-            pnr
 }
